Speed up (>2x) Huffman decoder by using a large table
The main insight of this optimization is that most of the time we deal
with data whose longest encoding no longer than 16 bits.  This allows
us to use a 64 KB table indexed by two bytes of input and to output up
to 3 bytes at a time.

In the case a longer code is encoutered, the fallback mechanism is to
use the original Huffman decoder that supports all codes.