Use the new XXH table; fix nameval lookup with empty value
Interestingly, this exercise resulted in finding and fixing a bug
in hash lookup for name and value match.  Before, if value was empty,
the value would never match.  (This was inherited from the original
code.)  Of course, just because a value is empty does not mean that
the header field could not be fully encoded using a static table
reference.